requirements.txt
streamlit
pandas
numpy
scipy
openai
langchain
langchain-community
matplotlib
paho-mqtt
faker
pytest
locust